whisper_server/requirements.txt
Flask==3.0.1
openai-whisper==20231117